[Upd] more rusty functions + updated db models

+++ src-tauri/migrations/2024-01-16-150430_create_receipts/up.sql
@@ -10,7 +10,7 @@ CREATE TABLE author (
 
 CREATE TABLE rating (
     id INTEGER NOT NULL PRIMARY KEY,
-    score INTEGER NOT NULL,  -- Average rating
+    score FLOAT NOT NULL,  -- Average rating
     amount INTEGER NOT NULL  -- Amount of votes
 );
